requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
idna==3.10
ImageHash==4.3.1
numpy==2.1.3
pillow==11.0.0
PyWavelets==1.7.0
requests==2.32.3
scipy==1.14.1
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
